+++ src/server/common/helpers/logging/request-logger.integration.test.js
@@ -1,28 +1,10 @@
-import hapi from '@hapi/hapi'
-import hapiPino from 'hapi-pino'
 import pino from 'pino'
 import { afterEach, describe, expect, it } from 'vitest'
 
 import { createLogger } from './logger.js'
 import { loggerOptions } from './logger-options.js'
 import { config } from '#config/config.js'
-
-const newServer = async () => {
-  const lines = []
-  const stream = { write: (s) => lines.push(s) }
-  const { transport: _transport, ...rest } = loggerOptions
-
-  const server = hapi.server({ port: 0 })
-  await server.register({
-    plugin: hapiPino,
-    options: { ...rest, enabled: true, level: 'trace', stream }
-  })
-
-  return { server, lines }
-}
-
-const findLine = (lines, predicate) =>
-  lines.map((s) => JSON.parse(s)).find(predicate)
+import { createLogCaptureServer } from '#server/common/test-helpers/log-capture-server.js'
 
 describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
   afterEach(() => {
@@ -31,7 +13,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should emit a canonical IndexedLogProperties payload through hapi-pino unchanged', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -45,7 +27,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'test event')
+    const out = findLine((l) => l.message === 'test event')
 
     expect(out).toMatchObject({
       message: 'test event',
@@ -56,7 +38,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should bind ECS http.* and url.* on the per-request child logger (no flat req)', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -67,7 +49,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'handler entered')
+    const out = findLine((l) => l.message === 'handler entered')
 
     expect(out).toMatchObject({
       http: { request: { method: 'GET' } },
@@ -79,7 +61,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should emit ECS http.response.* and url.path on the hapi-pino response auto-log', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -87,10 +69,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(
-      lines,
-      (l) => l.url?.path === '/test' && l.http?.response
-    )
+    const out = findLine((l) => l.url?.path === '/test' && l.http?.response)
 
     expect(out).toMatchObject({
       http: { response: { status_code: 200 } },
@@ -102,7 +81,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should map a logged err to ECS error.* when cdpEnvironment is non-prod', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -113,7 +92,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'boom path')
+    const out = findLine((l) => l.message === 'boom path')
 
     expect(out?.error).toStrictEqual(
       expect.objectContaining({
@@ -126,7 +105,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should strip error.stack_trace when cdpEnvironment is prod', async () => {
     config.set('cdpEnvironment', 'prod')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -137,7 +116,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'boom path')
+    const out = findLine((l) => l.message === 'boom path')
 
     expect(out?.error).toBeDefined()
     expect(out?.error).not.toHaveProperty('stack_trace')
