+++ src/server/common/helpers/request-tracing.integration.test.js
@@ -1,26 +1,16 @@
-import hapi from '@hapi/hapi'
-import hapiPino from 'hapi-pino'
 import { afterEach, describe, expect, it } from 'vitest'
 
-import { loggerOptions } from './logging/logger-options.js'
 import { requestTracing } from './request-tracing.js'
 import { config } from '#config/config.js'
+import { createLogCaptureServer } from '#server/common/test-helpers/log-capture-server.js'
 
 const UUID_V4 =
   /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i
 
 const newServer = async () => {
-  const lines = []
-  const stream = { write: (s) => lines.push(s) }
-  const { transport: _transport, ...rest } = loggerOptions
-
-  const server = hapi.server({ port: 0 })
-  await server.register({
-    plugin: hapiPino,
-    options: { ...rest, enabled: true, level: 'trace', stream }
-  })
-  await server.register(requestTracing)
-  server.route({
+  const captured = await createLogCaptureServer()
+  await captured.server.register(requestTracing)
+  captured.server.route({
     method: 'GET',
     path: '/test',
     handler: (request) => {
@@ -28,12 +18,8 @@ const newServer = async () => {
       return 'ok'
     }
   })
-
-  return { server, lines }
+  return captured
 }
-
-const findHandlerLine = (lines) =>
-  lines.map((s) => JSON.parse(s)).find((l) => l.message === 'inside handler')
 
 describe('#request-tracing', () => {
   describe('local trace.id fallback', () => {
@@ -43,48 +29,48 @@ describe('#request-tracing', () => {
 
     it('should mint a uuid trace.id when local and no x-cdp-request-id header is supplied', async () => {
       config.set('cdpEnvironment', 'local')
-      const { server, lines } = await newServer()
+      const { server, findLine } = await newServer()
 
       await server.inject('/test')
-      const out = findHandlerLine(lines)
+      const out = findLine((l) => l.message === 'inside handler')
 
       expect(out?.trace?.id).toMatch(UUID_V4)
     })
 
     it('should not mint a trace.id when not local and no x-cdp-request-id header is supplied', async () => {
       config.set('cdpEnvironment', 'dev')
-      const { server, lines } = await newServer()
+      const { server, findLine } = await newServer()
 
       await server.inject('/test')
-      const out = findHandlerLine(lines)
+      const out = findLine((l) => l.message === 'inside handler')
 
       expect(out).not.toHaveProperty('trace')
     })
 
     it('should pass through an incoming x-cdp-request-id header as trace.id regardless of environment', async () => {
       config.set('cdpEnvironment', 'prod')
-      const { server, lines } = await newServer()
+      const { server, findLine } = await newServer()
 
       await server.inject({
         method: 'GET',
         url: '/test',
         headers: { 'x-cdp-request-id': 'incoming-trace-abc' }
       })
-      const out = findHandlerLine(lines)
+      const out = findLine((l) => l.message === 'inside handler')
 
       expect(out?.trace?.id).toBe('incoming-trace-abc')
     })
 
     it('should not overwrite an incoming x-cdp-request-id header when local', async () => {
       config.set('cdpEnvironment', 'local')
-      const { server, lines } = await newServer()
+      const { server, findLine } = await newServer()
 
       await server.inject({
         method: 'GET',
         url: '/test',
         headers: { 'x-cdp-request-id': 'incoming-trace-xyz' }
       })
-      const out = findHandlerLine(lines)
+      const out = findLine((l) => l.message === 'inside handler')
 
       expect(out?.trace?.id).toBe('incoming-trace-xyz')
     })
